move(cosmo): move units to private API
